+++ src/Firebase/Messaging/ApnsConfig.php
@@ -6,13 +6,21 @@ namespace Kreait\Firebase\Messaging;
 
 use JsonSerializable;
 
+/**
+ * @see https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/generating_a_remote_notification
+ */
 final class ApnsConfig implements JsonSerializable
 {
     /** @var array<string, mixed> */
-    private $rawConfig;
+    private $config;
 
     private function __construct()
     {
+    }
+
+    public static function new(): self
+    {
+        return self::fromArray([]);
     }
 
     /**
@@ -21,7 +29,40 @@ final class ApnsConfig implements JsonSerializable
     public static function fromArray(array $data): self
     {
         $config = new self();
-        $config->rawConfig = $data;
+        $config->config = $data;
+
+        return $config;
+    }
+
+    public function withDefaultSound(): self
+    {
+        return $this->withSound('default');
+    }
+
+    /**
+     * The name of a sound file in your app’s main bundle or in the Library/Sounds folder of your app’s
+     * container directory. Specify the string "default" to play the system sound.
+     */
+    public function withSound(string $sound): self
+    {
+        $config = clone $this;
+
+        $config->config['payload'] = $config->config['payload'] ?? [];
+        $config->config['payload']['aps'] = $config->config['payload']['aps'] ?? [];
+        $config->config['payload']['aps']['sound'] = $sound;
+
+        return $config;
+    }
+
+    /**
+     * The number to display in a badge on your app’s icon. Specify 0 to remove the current badge, if any.
+     */
+    public function withBadge(int $number): self
+    {
+        $config = clone $this;
+        $config->config['payload'] = $config->config['payload'] ?? [];
+        $config->config['payload']['aps'] = $config->config['payload']['aps'] ?? [];
+        $config->config['payload']['aps']['badge'] = $number;
 
         return $config;
     }
@@ -31,7 +72,7 @@ final class ApnsConfig implements JsonSerializable
      */
     public function jsonSerialize(): array
     {
-        return \array_filter($this->rawConfig, static function ($value) {
+        return \array_filter($this->config, static function ($value) {
             return $value !== null;
         });
     }